stocks_products/requirements.txt
asgiref==3.6.0
Django==4.1.4
django-filter==22.1
djangorestframework==3.14.0
# psycopg2-binary==2.9.5
# python-dotenv==0.21.1
pytz==2022.7
sqlparse==0.4.3
